feature(save): Add save as option in edit menu.
This new feature allows you to save files on disk, it shows an alert when no file has been opened and it does not change the build code.

This implementation works, but is relying on the "active" class used for the tabs in IDE and some other css asumptions, along with the localStorage file name standard used by the original remix dependency. Also, it is needed to keep tracking of saved files in some way. This implementation is a first approach.